CustomPath v0.2 - bug first path add after second drawn

+++ app/src/main/java/il/ghostdog/drawingapp/Vector2.kt
@@ -1,4 +1,4 @@
 package il.ghostdog.drawingapp
 
 @kotlinx.serialization.Serializable
-data class Vector2(var x: Float, var y: Float)
+data class Vector2(var x: Int, var y: Int)
